requirements.txt
python-telegram-bot==13.1
transmissionrpc==0.11
